+++ its_live_monitoring/src/constants.py
@@ -1,4 +1,0 @@
-"""Constants that are shared between multiple files."""
-
-MAX_PAIR_SEPARATION_IN_DAYS = 544
-MAX_CLOUD_COVER_PERCENT = 60