+++ tests/conftest.py
@@ -16,6 +16,7 @@ def pystac_item_factory():
         collection: str,
         geometry: dict | None = None,
         bbox: list | None = None,
+        assets: dict = None,
     ) -> pystac.item.Item:
         if isinstance(datetime, str):
             datetime = date_parser(datetime)
@@ -27,6 +28,7 @@ def pystac_item_factory():
             datetime=datetime,
             properties=properties,
             collection=collection,
+            assets=assets,
         )
 
         return expected_item